requirements.txt
# ctransformers==0.2.27
sentence-transformers==2.2.2
pinecone-client==3.2.1
langchain==0.1.13
flask
pypdf
-e .
python-dotenv
einops
torch